calorie-counter/requirements.txt
openai
flask
python-dotenv
gunicron
mysql-connector-python
flask-sqlalchemy
requests
beautifulsoup4
lxml
